Fix hang on netdde crash with a nacked interrupt
If netdde crashed without acknowledging the interrupt, the hardware
might be stuck in a situation where it keeps the interrupt raised, and
thus linux_intr keeps getting called. linux_intr should detect that we
have removed handlers and then just disable the interrupt. Hopefully
when the driver restarts it will reset the card, and thus the interrupt
will stop before it enables it again.